+++ apps/extension/src/ui/apps/dashboard/routes/Portfolio/PortfolioAssets.tsx
@@ -1,9 +1,9 @@
 import { Balances } from "@core/domains/balances/types"
 import { DashboardAssetsTable } from "@ui/domains/Portfolio/AssetsTable"
-import { usePortfolio } from "@ui/domains/Portfolio/context"
 import { NetworkPicker } from "@ui/domains/Portfolio/NetworkPicker"
 import { Statistics } from "@ui/domains/Portfolio/Statistics"
 import { useDisplayBalances } from "@ui/domains/Portfolio/useDisplayBalances"
+import { usePortfolio } from "@ui/domains/Portfolio/usePortfolio"
 import { useAnalytics } from "@ui/hooks/useAnalytics"
 import { useSelectedCurrency } from "@ui/hooks/useCurrency"
 import { useHasAccounts } from "@ui/hooks/useHasAccounts"
