+++ apps/extension/src/ui/apps/popup/pages/Portfolio/PortfolioAsset.tsx
@@ -3,9 +3,9 @@ import { ChevronLeftIcon } from "@talismn/icons"
 import { Fiat } from "@ui/domains/Asset/Fiat"
 import { TokenLogo } from "@ui/domains/Asset/TokenLogo"
 import { PopupAssetDetails } from "@ui/domains/Portfolio/AssetDetails"
-import { usePortfolio } from "@ui/domains/Portfolio/context"
-import { useSelectedAccount } from "@ui/domains/Portfolio/SelectedAccountContext"
 import { useDisplayBalances } from "@ui/domains/Portfolio/useDisplayBalances"
+import { usePortfolio } from "@ui/domains/Portfolio/usePortfolio"
+import { useSelectedAccount } from "@ui/domains/Portfolio/useSelectedAccount"
 import { useTokenBalancesSummary } from "@ui/domains/Portfolio/useTokenBalancesSummary"
 import { useAnalytics } from "@ui/hooks/useAnalytics"
 import useBalances from "@ui/hooks/useBalances"
