+++ apps/extension/src/ui/domains/Portfolio/NetworkPicker.tsx
@@ -2,7 +2,7 @@ import { Combobox } from "@headlessui/react"
 import { ChevronDownIcon, SearchIcon, XIcon } from "@talismn/icons"
 import { classNames } from "@talismn/util"
 import { ChainLogo } from "@ui/domains/Asset/ChainLogo"
-import { NetworkOption, usePortfolio } from "@ui/domains/Portfolio/context"
+import { NetworkOption, usePortfolio } from "@ui/domains/Portfolio/usePortfolio"
 import { useCallback, useEffect, useMemo, useRef, useState } from "react"
 import { useTranslation } from "react-i18next"
 import { IconButton } from "talisman-ui"
