+++ apps/extension/src/ui/domains/Portfolio/NoTokensMessage.tsx
@@ -1,5 +1,5 @@
 import { ArrowDownIcon, CreditCardIcon } from "@talismn/icons"
-import { useSelectedAccount } from "@ui/domains/Portfolio/SelectedAccountContext"
+import { useSelectedAccount } from "@ui/domains/Portfolio/useSelectedAccount"
 import { useAnalytics } from "@ui/hooks/useAnalytics"
 import { useIsFeatureEnabled } from "@ui/hooks/useFeatures"
 import { useCallback } from "react"
