+++ apps/extension/src/ui/domains/Portfolio/useTokenBalancesSummary.ts
@@ -2,7 +2,7 @@ import { Balances } from "@core/domains/balances/types"
 import { Chain } from "@core/domains/chains/types"
 import { Token } from "@core/domains/tokens/types"
 import { TokenRatesList } from "@talismn/token-rates"
-import { usePortfolio } from "@ui/domains/Portfolio/context"
+import { usePortfolio } from "@ui/domains/Portfolio/usePortfolio"
 import { useSelectedCurrency } from "@ui/hooks/useCurrency"
 import BigNumber from "bignumber.js"
 import { useMemo } from "react"
